amlbian: add T7 for supportting debian [1/1]
PD#SWPL-106598

Problem:
add T7 for supportting debian system.

Solution:
add new kernel deconfig file.

Verify:
local

+++ drivers/usb/gadget/v3_burning/v3_common/v3_tool_buff_manager.c
@@ -84,7 +84,7 @@ int v3tool_buffman_img_verify_sha1sum(unsigned char* vrySum)
         {
             case V3TOOL_PART_IMG_FMT_RAW:
                 {
-                    int thisVryLen = 0;
+					unsigned long thisVryLen = 0;
                     static unsigned long _lastReportTick = 0;
                     if (!vryLen) _lastReportTick = get_timer(0);
                     for (; vryLen < imgTotalLen; vryLen += thisVryLen)
